configuration tile
